refactor(document): 重构文档策略确认功能实现
- 添加ConfirmStrategy聚合根用于封装策略确认的事务性操作
- 实现SaveConfirmStrategyAggregate方法支持原子性保存多个实体变更
- 更新ConfirmStrategy逻辑返回文档和策略方案两个实体对象
- 移除buildStrategySnapshot方法，将其功能集成到FillAndProcessPipeline中
- 修复DocumentStrategyStep模型中PlanId字段类型从int改为int64
- 添加FromConfirmStrategyReq和ToConfirmStrategyResp转换器方法
- 完善策略确认过程中的状态校验和日志记录机制

+++ internal/infrastructure/model/document_strategy.go
@@ -25,7 +25,7 @@ type DocumentStrategyPlan struct {
 type DocumentStrategyStep struct {
 	common.Model
 	DocumentId      int64  `gorm:"column:document_id"`      // 文档ID
-	PlanId          int    `gorm:"column:plan_id"`          // 方案ID
+	PlanId          int64  `gorm:"column:plan_id"`          // 方案ID
 	StepNo          int    `gorm:"column:step_no"`          // 步骤序号
 	PipelineType    int    `gorm:"column:pipeline_type"`    // 流水线类型
 	StrategyType    int    `gorm:"column:strategy_type"`    // 策略类型
